Change 2561711 by Nick.Darnell@Nick.Darnell_Dev on 2015/05/21 18:18:13
Slate - Attempting to find cases where we're releasing UTextures used in slate by allowing them to be garbage collected, but then attempting to use them later in slate with free'ed memory pointed to by the slate brush.  When this happens we ensure and use a fallback texture instead in non-shipping builds.

[CL 2581961 by Peter Knepley in UT-Clean branch]